Fix misassigning of accessToken & log more
At one point, access_token was  set to string "null", which is what `.as<String>()` will do if missing from the document. Unknown why it was missing from the document initially, as the log rotated out very quickly due to the expires_in also being missing, which meant that it was getting set to 0 and causing token refreshes on every loop.

Hopefully with the guards and logging added, if this happens in the future, we'll be able to figure out what happened.