Don't restart an album that's already playing
Allows you to pick up the record while it's playing to look at it without causing it to start over again from the beginning when you put it back down.

We start getting connection-related errors after running for several hours, so we add some extra logging around SSL.

Turns out HTTPClient will handle Content-Length for us, so we remove it.